mock airbyte response
