+++ ddpui/tests/unit_integration_tests/test_airbyte_service.py
@@ -1,4 +1,5 @@
 import os
+from unittest import mock
 from unittest.mock import patch
 import django
 from pydantic import ValidationError
@@ -8,6 +9,7 @@ os.environ.setdefault("DJANGO_SETTINGS_MODULE", "ddpui.settings")
 os.environ["DJANGO_ALLOW_ASYNC_UNSAFE"] = "true"
 django.setup()
 
+from ninja.errors import HttpError
 from ddpui.tests.helper.test_airbyte_unit_schemas import *
 from ddpui.ddpairbyte.airbyte_service import *
 
@@ -31,7 +33,7 @@ def valid_name():
 
 @pytest.fixture
 def invalid_name():
-    return
+    return 123
 
 
 @pytest.fixture(scope="module")
@@ -46,526 +48,566 @@ def valid_sourcedef_id(valid_workspace_id):
 
 
 def mock_abreq(endpoint, data):
-    return {}
+    return {"connectionSpecification": {"test": "data"}}
 
 
 def test_abreq_success():
     endpoint = "workspaces/list"
-    result = abreq(endpoint)
+    expected_response = {
+        "workspaces": [{"workspaceId": "1", "name": "Example Workspace"}]
+    }
+
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 200
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = expected_response
+
+        result = abreq(endpoint)
+
     assert isinstance(result, dict)
+    assert result == expected_response
     assert "workspaces" in result
     assert isinstance(result["workspaces"], list)
 
 
-def test_abreq_invalid_endpoint():
-    endpoint = "invalid_endpoint"
-    with pytest.raises(HttpError) as excinfo:
-        abreq(endpoint)
-    assert str(excinfo.value).startswith("Something went wrong: ('404 Client Error:")
-
-
-def test_abreq_invalid_request_data():
-    endpoint = "workspaces/create"
-    req = {"invalid_key": "invalid_value"}
-    with pytest.raises(HttpError) as excinfo:
-        abreq(endpoint, req)
-    assert str(excinfo.value).startswith("Something went wrong: ('500 Server Error:")
-
-
 def test_abreq_connection_error():
-    os.environ["AIRBYTE_SERVER_HOST"] = "invalid_host"
-    with pytest.raises(HttpError) as excinfo:
-        abreq("test_endpoint", {"test": "data"})
+    endpoint = "my_endpoint"
 
-    assert str(excinfo.value) == "Error connecting to Airbyte server"
-    os.environ["AIRBYTE_SERVER_HOST"] = "localhost"
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.side_effect = requests.exceptions.ConnectionError()
+
+        with pytest.raises(HttpError) as excinfo:
+            abreq(endpoint)
+
+        assert excinfo.value.status_code == 500
+        assert str(excinfo.value) == "Error connecting to Airbyte server"
+
+
+# def test_abreq_invalid_request_data():
+#     endpoint = "workspaces/create"
+#     req = {"invalid_key": "invalid_value"}
+
+#     with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+#         mock_post.return_value.status_code = 400
+#         mock_post.return_value.headers = {"Content-Type": "application/json"}
+#         mock_post.return_value.json.return_value = {"error": "Invalid request data"}
+#         with pytest.raises(HttpError) as excinfo:
+#             abreq(endpoint, req)
+#         assert excinfo.value.status_code == 400
+#         assert str(excinfo.value) == "Something went wrong: Invalid request data"
 
 
 def test_get_workspaces_success():
-    result = get_workspaces()
-    assert isinstance(result, list)
-    assert all(isinstance(workspace, dict) for workspace in result)
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 200
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {
+            "workspaces": [{"workspaceId": "1", "name": "Example Workspace"}]
+        }
+
+        result = get_workspaces()
+        assert isinstance(result, list)
+        assert all(isinstance(workspace, dict) for workspace in result)
+
+
+def test_get_workspaces_failure():
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 404
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {"error": "Invalid request data"}
+        with pytest.raises(HttpError) as excinfo:
+            get_workspaces()
+        assert excinfo.value.status_code == 404
+        assert str(excinfo.value) == "no workspaces found"
 
 
 def test_create_workspace_with_valid_name(valid_name):
-    result = create_workspace(valid_name)
-    assert "workspaceId" in result
-    assert isinstance(result, dict)
+    # check if workspace is created successfully using mock_abreq
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 200
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {
+            "workspaceId": "1",
+            "name": "Example Workspace",
+        }
+
+        result = create_workspace(valid_name)
+        assert "workspaceId" in result
+        assert isinstance(result, dict)
 
 
 def test_create_workspace_invalid_name():
-    with pytest.raises(TypeError) as excinfo:
+    with pytest.raises(HttpError) as excinfo:
         create_workspace(123)
     assert str(excinfo.value) == "Name must be a string"
 
 
 def test_create_workspace_failure():
-    with patch("ddpui.ddpairbyte.airbyte_service.abreq", mock_abreq):
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 400
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {"error": "Invalid request data"}
         with pytest.raises(HttpError) as excinfo:
             create_workspace("test_workspace")
+        assert excinfo.value.status_code == 400
         assert str(excinfo.value) == "workspace not created"
 
 
-def test_get_workspace_success(valid_workspace_id):
-    result = get_workspace(valid_workspace_id)
-    assert "workspaceId" in result
-    assert isinstance(result, dict)
+def test_get_workspace_success():
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 200
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {
+            "workspaceId": "test",
+            "name": "Example Workspace",
+        }
+        result = get_workspace("test")
+        assert "workspaceId" in result
+        assert isinstance(result, dict)
 
 
-def test_get_workspace_failure():
-    with pytest.raises(TypeError) as excinfo:
+def test_get_workspace_success_with_invalid_workspace_id():
+    with pytest.raises(HttpError) as excinfo:
         get_workspace(123)
     assert str(excinfo.value) == "workspace_id must be a string"
 
 
-def test_get_workspace_with_invalid_workspace_id():
-    with patch("ddpui.ddpairbyte.airbyte_service.abreq", mock_abreq):
+def test_get_workspace_failure():
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 404
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {"error": "Invalid request data"}
         with pytest.raises(HttpError) as excinfo:
             get_workspace("test")
+        assert excinfo.value.status_code == 404
         assert str(excinfo.value) == "workspace not found"
 
 
-def test_set_workspace_name_success(valid_workspace_id, valid_name):
-    result = set_workspace_name(valid_workspace_id, valid_name)
-    assert "workspaceId" in result
-    assert isinstance(result, dict)
+def test_set_workspace_name_success():
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 200
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {
+            "workspaceId": "test",
+            "name": "Example Workspace",
+        }
+        result = set_workspace_name("test", "New Name")
+        assert "workspaceId" in result
+        assert isinstance(result, dict)
 
 
-def test_set_workspace_name_failure():
-    with pytest.raises(TypeError) as excinfo:
-        set_workspace_name(123, "test_name")
+def test_set_workspace_name_with_invalid_workspace_id():
+    with pytest.raises(HttpError) as excinfo:
+        set_workspace_name(123, "New Name")
     assert str(excinfo.value) == "Workspace ID must be a string"
 
 
 def test_set_workspace_name_with_invalid_name():
-    with pytest.raises(TypeError) as excinfo:
-        set_workspace_name("123", 123)
+    with pytest.raises(HttpError) as excinfo:
+        set_workspace_name("test", 123)
     assert str(excinfo.value) == "Name must be a string"
 
 
-def test_set_workspace_name_with_invalid_workspace_id():
-    with patch("ddpui.ddpairbyte.airbyte_service.abreq", mock_abreq):
+def test_set_workspace_name_failure():
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 404
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {"error": "Invalid request data"}
         with pytest.raises(HttpError) as excinfo:
-            set_workspace_name("test", "test")
-    assert str(excinfo.value) == "workspace not found"
+            set_workspace_name("test", "New Name")
+        assert excinfo.value.status_code == 404
+        assert str(excinfo.value) == "workspace not found"
 
 
-class TestDeleteSource:
-    def test_create_workspace(self):
-        """creates a workspace, checks airbyte response"""
-        payload = {"name": "test_workspace"}
-
-        try:
-            CreateWorkspaceTestPayload(**payload)
-        except ValidationError as error:
-            raise ValueError(
-                f"Field do not match in the payload: {error.errors()}"
-            ) from error
-
-        try:
-            res = create_workspace(**payload)
-            CreateWorkspaceTestResponse(**res)
-            TestWorkspace.workspace_id = res["workspaceId"]
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
-
-    def test_a_create_source(self, test_workspace_id):
-        source_definitions = get_source_definitions(workspace_id=test_workspace_id)
-        for source_definition in source_definitions:
-            if source_definition["name"] == "File (CSV, JSON, Excel, Feather, Parquet)":
-                source_definition_id = source_definition["sourceDefinitionId"]
-                break
-
-        payload = {
-            "sourcedef_id": source_definition_id,
-            "config": {
-                "url": "https://storage.googleapis.com/covid19-open-data/v2/latest/epidemiology.csv",
-                "format": "csv",
-                "provider": {"storage": "HTTPS"},
-                "dataset_name": "covid19data",
-            },
-            "workspace_id": str(test_workspace_id),
-            "name": "source1",
-        }
-        try:
-            CreateSourceTestPayload(**payload)
-        except ValidationError as e:
-            raise ValueError(f"Field do not match in payload: {e.errors()}")
-        try:
-            res = create_source(**payload)
-            CreateSourceTestResponse(**res)
-            TestAirbyteSource.source_id = res["sourceId"]
-            TestAirbyteSource.source_definition_id = res["sourceDefinitionId"]
-        except ValidationError as e:
-            raise ValueError(f"Response validation failed: {e.errors()}")
-
-
-class TestWorkspace:
-    """class which holds all the workspace tests"""
-
-    workspace_id = None
-
-    def test_create_workspace(self):
-        """creates a workspace, checks airbyte response"""
-        payload = {"name": "test_workspace"}
-
-        try:
-            CreateWorkspaceTestPayload(**payload)
-        except ValidationError as error:
-            raise ValueError(
-                f"Field do not match in the payload: {error.errors()}"
-            ) from error
-
-        try:
-            res = create_workspace(**payload)
-            CreateWorkspaceTestResponse(**res)
-            TestWorkspace.workspace_id = res["workspaceId"]
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
-
-    def test_get_workspace(self):
-        """gets a workspace, checks airbyte response"""
-        try:
-            res = get_workspace(workspace_id=TestWorkspace.workspace_id)
-            GetWorkspaceTestResponse(**res)
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
-
-    def test_get_workspaces(self):
-        """gets all workspaces, checks airbyte response"""
-        try:
-            res = get_workspaces()
-            GetWorkspacesTestResponse(**res)
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
-
-    def test_set_workspace_name(self):
-        """sets workspace name, checks airbyte response"""
-        new_name = "test"
-
-        try:
-            res = set_workspace_name(
-                workspace_id=TestWorkspace.workspace_id, name=new_name
-            )
-            SetWorkspaceTestResponse(**res)
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
-
-
-@pytest.fixture(scope="session")
-def test_workspace_id():
-    """workspace id of test workspace"""
-    test_workspace = TestWorkspace()
-    test_workspace.test_create_workspace()
-    return test_workspace.workspace_id
-
-
-class TestAirbyteSource:
-    """class which holds all the source tests"""
-
-    source_config = {
-        "url": "https://storage.googleapis.com/covid19-open-data/v2/latest/epidemiology.csv",
-        "format": "csv",
-        "provider": {"storage": "HTTPS"},
-        "dataset_name": "covid19data",
-    }
-
-    def test_source_connection(self, test_workspace_id):
-        """tests connectivity to a source"""
-        source_definitions = get_source_definitions(workspace_id=test_workspace_id)
-        for source_definition in source_definitions:
-            if source_definition["name"] == "File (CSV, JSON, Excel, Feather, Parquet)":
-                TestAirbyteSource.source_definition_id = source_definition[
-                    "sourceDefinitionId"
-                ]
-                break
-
-        try:
-            res = check_source_connection(
-                test_workspace_id,
-                AirbyteSourceCreate(
-                    name="unused",
-                    sourceDefId=TestAirbyteSource.source_definition_id,
-                    config=self.source_config,
-                ),
-            )
-            CheckSourceConnectionTestResponse(**res)
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
-
-    def test_a_create_source(self, test_workspace_id):
-        """tests source creation"""
-        payload = {
-            "sourcedef_id": TestAirbyteSource.source_definition_id,
-            "config": TestAirbyteSource.source_config,
-            "workspace_id": str(test_workspace_id),
-            "name": "source1",
-        }
-        try:
-            CreateSourceTestPayload(**payload)
-        except ValidationError as error:
-            raise ValueError(
-                f"Field do not match in payload: {error.errors()}"
-            ) from error
-        try:
-            res = create_source(**payload)
-            CreateSourceTestResponse(**res)
-            TestAirbyteSource.source_id = res["sourceId"]
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
-
-    def test_get_definitions(self, test_workspace_id):
-        """tests retrieval of source definitions"""
-        try:
-            res = get_source_definitions(workspace_id=test_workspace_id)
-            GetSourceDefinitionsTestResponse(__root__=res)
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
-
-    def test_get_source_schema_catalog(self, test_workspace_id):
-        """fetches the schema catalog for a source"""
-        try:
-            res = get_source_schema_catalog(
-                test_workspace_id, TestAirbyteSource.source_id
-            )
-            GetSourceSchemaCatalogTestResponse(catalog=res)
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
-
-    def test_fail_to_get_source_schema_catalog(self, test_workspace_id):
-        """fetches the schema catalog for a source"""
-        try:
-            get_source_schema_catalog(test_workspace_id, "not-a-source-id")
-        except AirbyteError:
-            pass
-
-    def test_get_source(self, test_workspace_id):
-        """tests retrieval of a single source"""
-        try:
-            res = get_source(
-                workspace_id=test_workspace_id,
-                source_id=TestAirbyteSource.source_id,
-            )
-            GetSourceTestResponse(**res)
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
-
-    def test_get_sources(self, test_workspace_id):
-        """tests retrieval of all sources"""
-        try:
-            res = get_sources(workspace_id=test_workspace_id)
-            GetSourcesTestResponse(sources=res)
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
-
-    def test_update_source(self):
-        """tests updating a source"""
-        payload = {
-            "name": "source9",
-            "sourcedef_id": TestAirbyteSource.source_definition_id,
-            "source_id": TestAirbyteSource.source_id,
-            "config": {
-                "url": "https://storage.googleapis.com/covid19-open-data/v2/latest/epidemiology.csv",
-                "format": "csv",
-                "provider": {"storage": "HTTPS"},
-                "dataset_name": "covid19data",
-            },
-        }
-        try:
-            UpdateSourceTestPayload(**payload)
-        except ValidationError as error:
-            raise ValueError(
-                f"Field do not match in payload: {error.errors()}"
-            ) from error
-
-        try:
-            res = update_source(**payload)
-            UpdateSourceTestResponse(**res)
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
-
-
-@pytest.fixture(scope="session")
-def test_source_id(test_workspace_id):
-    test_source = TestAirbyteSource()
-    test_source.test_a_create_source(test_workspace_id)
-    return test_source.source_id
-
-
-class TestAirbyteDestination:
-    destination_config = {
-        "host": os.getenv("DBHOST"),
-        "port": 5432,
-        "database": os.getenv("DBNAME"),
-        "username": os.getenv("DBUSER"),
-        "schema": "staging",
-    }
-
-    def test_a_create_destination(self, test_workspace_id):
-        destination_definitions = get_destination_definitions(
-            workspace_id=test_workspace_id
-        )
-        for destination_definition in destination_definitions:
-            if destination_definition["name"] == "Postgres":
-                destination_definition_id = destination_definition[
-                    "destinationDefinitionId"
-                ]
-                break
-
-        TestAirbyteDestination.destination_definition_id = destination_definition_id
-        payload = {
-            "name": "destination1",
-            "destinationdef_id": TestAirbyteDestination.destination_definition_id,
-            "config": TestAirbyteDestination.destination_config,
-            "workspace_id": str(test_workspace_id),
-        }
-        try:
-            CreateDestinationTestPayload(**payload)
-        except ValidationError as error:
-            raise ValueError(
-                f"Field do not match in payload: {error.errors()}"
-            ) from error
-
-        try:
-            res = create_destination(**payload)
-            CreateDestinationTestResponse(**res)
-            TestAirbyteDestination.destination_id = res["destinationId"]
-            TestAirbyteDestination.destination_definition_id = res[
-                "destinationDefinitionId"
+def test_get_source_definitions_success():
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 200
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {
+            "sourceDefinitions": [
+                {"sourceDefinitionId": "1", "name": "Example Source Definition 1"},
+                {"sourceDefinitionId": "2", "name": "Example Source Definition 2"},
             ]
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
-
-    def test_get_destination(self, test_workspace_id):
-        try:
-            res = get_destination(
-                workspace_id=test_workspace_id,
-                destination_id=TestAirbyteDestination.destination_id,
-            )
-            GetDestinationTestResponse(**res)
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
-
-    def test_update_destination(self):
-        payload = {
-            "name": "test_update",
-            "destination_id": TestAirbyteDestination.destination_id,
-            "config": {
-                "host": os.getenv("DBHOST"),
-                "port": 5432,
-                "database": os.getenv("DBNAME"),
-                "username": os.getenv("DBUSER"),
-                "schema": "staging",
-            },
-            "destinationdef_id": TestAirbyteDestination.destination_definition_id,
         }
-
-        try:
-            UpdateDestinationTestPayload(**payload)
-        except ValidationError as error:
-            raise ValueError(
-                f"Field do not match in payload: {error.errors()}"
-            ) from error
-
-        try:
-            res = update_destination(**payload)
-            UpdateDestinationTestResponse(**res)
-        except ValidationError as error:
-            raise ValueError(
-                f"Field do not match in resposne: {error.errors()}"
-            ) from error
-
-    def test_check_destination_connection(self, test_workspace_id):
-        workspace_id = test_workspace_id
-        try:
-            res = check_destination_connection(
-                workspace_id,
-                AirbyteDestinationCreate(
-                    name="unused",
-                    destinationDefId=TestAirbyteDestination.destination_definition_id,
-                    config=TestAirbyteDestination.destination_config,
-                ),
-            )
-            CheckDestinationConnectionTestResponse(**res)
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
+        result = get_source_definitions("test")
+        assert isinstance(result, list)
 
 
-@pytest.fixture(scope="session")
-def test_destination_id(test_workspace_id):
-    test_destination = TestAirbyteDestination()
-    test_destination.test_a_create_destination(test_workspace_id)
-    return test_destination.destination_id
+def test_get_source_definitions_failure():
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 404
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {"error": "Invalid request data"}
+        with pytest.raises(HttpError) as excinfo:
+            get_source_definitions("test")
+        assert excinfo.value.status_code == 404
+        assert str(excinfo.value) == "source definitions not found for workspace"
 
 
-class TestConnection:
-    def test_a_create_connection(
-        self,
-        test_workspace_id,
-        test_source_id,
-        test_destination_id,
-    ):
-        workspace_id = str(test_workspace_id)
-        connection_info = schema.AirbyteConnectionCreate(
-            sourceId=str(test_source_id),
-            destinationId=str(test_destination_id),
-            name="Test Connection",
-            streams=[
-                {
-                    "name": "covid19data",
-                    "selected": True,
-                    "syncMode": "full_refresh",
-                    "destinationSyncMode": "overwrite",
-                }
-            ],
-            normalize=False,
-        )
+def test_get_source_definitions_with_invalid_workspace_id():
+    with pytest.raises(HttpError) as excinfo:
+        get_source_definitions(123)
+    assert str(excinfo.value) == "Invalid workspace ID"
 
-        test_airbyte_norm_op_id = create_normalization_operation(test_workspace_id)
-        try:
-            res = create_connection(
-                workspace_id, test_airbyte_norm_op_id, connection_info
-            )
-            CreateConnectionTestResponse(**res)
-            TestConnection.connection_id = res["connectionId"]
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
 
-    def test_get_connection(self, test_workspace_id):
-        workspace_id = test_workspace_id
-        connection_id = TestConnection.connection_id
+def test_get_source_definition_specification_success():
+    workspace_id = "my_workspace_id"
+    sourcedef_id = "my_sourcedef_id"
+    expected_response = {"key": "value"}
 
-        try:
-            res = get_connection(workspace_id, connection_id)
-            GetConnectionTestResponse(**res)
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
+    with patch("ddpui.ddpairbyte.airbyte_service.abreq") as mock_abreq:
+        mock_abreq.return_value = {"connectionSpecification": expected_response}
+        result = get_source_definition_specification(workspace_id, sourcedef_id)
 
-    def test_update_connection(
-        self, test_workspace_id, test_source_id, test_destination_id
-    ):
-        workspace_id = test_workspace_id
-        connection_id = (TestConnection.connection_id,)
-        connection_info = schema.AirbyteConnectionUpdate(
-            sourceId=test_source_id,
-            destinationId=test_destination_id,
-            connectionId=TestConnection.connection_id,
-            streams=[{"name": "companies"}],
-            name="Test Connection",
-        )
+    assert result == expected_response
+    assert isinstance(result, dict)
 
-        try:
-            res = update_connection(workspace_id, connection_id, connection_info)
-            UpdateConnectionTestResponse(**res)
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
 
-    def test_delete_connection(self, test_workspace_id):
-        workspace_id = test_workspace_id
-        connection_id = TestConnection.connection_id
+def test_get_source_definition_specification_failure():
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 404
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {"error": "Invalid request data"}
+        with pytest.raises(HttpError) as excinfo:
+            get_source_definition_specification("test", "1")
+        assert excinfo.value.status_code == 404
+        assert str(excinfo.value) == "specification not found for source definition"
 
-        try:
-            res = delete_connection(workspace_id, connection_id)
-            assert res == {}
-        except ValidationError as error:
-            raise ValueError(f"Response validation failed: {error.errors()}") from error
+
+def test_get_source_definition_specification_with_invalid_workspace_id():
+    with pytest.raises(HttpError) as excinfo:
+        get_source_definition_specification(123, "1")
+    assert str(excinfo.value) == "Invalid workspace ID"
+
+
+def test_get_source_definition_specification_with_invalid_source_definition_id():
+    with pytest.raises(HttpError) as excinfo:
+        get_source_definition_specification("test", 123)
+    assert str(excinfo.value) == "Invalid source definition ID"
+
+
+def test_get_sources_success():
+    workspace_id = "my_workspace_id"
+    expected_response = {"sources": [{"sourceId": "1", "name": "Example Source 1"}]}
+
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 200
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = expected_response
+        result = get_sources(workspace_id)
+
+        assert isinstance(result, list)
+
+
+def test_get_sources_failure():
+    workspace_id = "my_workspace_id"
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 404
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {"error": "Invalid request data"}
+        with pytest.raises(HttpError) as excinfo:
+            get_sources(workspace_id)
+        assert excinfo.value.status_code == 404
+        assert str(excinfo.value) == "sources not found for workspace"
+
+
+def test_get_sources_with_invalid_workspace_id():
+    with pytest.raises(HttpError) as excinfo:
+        get_sources(123)
+    assert str(excinfo.value) == "Invalid workspace ID"
+
+
+def test_get_source_success():
+    workspace_id = "my_workspace_id"
+    source_id = "1"
+    expected_response = {"sourceId": "1", "name": "Example Source 1"}
+
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 200
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = expected_response
+        result = get_source(workspace_id, source_id)
+
+        assert result == expected_response
+        assert isinstance(result, dict)
+
+
+def test_get_source_failure():
+    workspace_id = "my_workspace_id"
+    source_id = "1"
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 404
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {"error": "Invalid request data"}
+        with pytest.raises(HttpError) as excinfo:
+            get_source(workspace_id, source_id)
+        assert excinfo.value.status_code == 404
+        assert str(excinfo.value) == "source not found"
+
+
+def test_get_source_with_invalid_workspace_id():
+    with pytest.raises(HttpError) as excinfo:
+        get_source(123, "1")
+    assert str(excinfo.value) == "Invalid workspace ID"
+
+
+def test_get_source_with_invalid_source_id():
+    with pytest.raises(HttpError) as excinfo:
+        get_source("test", 123)
+    assert str(excinfo.value) == "Invalid source ID"
+
+
+def test_delete_source_success():
+    workspace_id = "my_workspace_id"
+    source_id = "1"
+    expected_response = {"sourceId": "1", "name": "Example Source 1"}
+
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 200
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = expected_response
+        result = delete_source(workspace_id, source_id)
+
+        assert result == expected_response
+        assert isinstance(result, dict)
+
+
+def test_delete_source_failure():
+    workspace_id = "my_workspace_id"
+    source_id = "1"
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 404
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {"error": "Invalid request data"}
+        with pytest.raises(HttpError) as excinfo:
+            delete_source(workspace_id, source_id)
+        assert excinfo.value.status_code == 404
+        assert str(excinfo.value) == "source not found"
+
+
+def test_delete_source_with_invalid_workspace_id():
+    with pytest.raises(HttpError) as excinfo:
+        delete_source(123, "1")
+    assert str(excinfo.value) == "Invalid workspace ID"
+
+
+def test_delete_source_with_invalid_source_id():
+    with pytest.raises(HttpError) as excinfo:
+        delete_source("test", 123)
+    assert str(excinfo.value) == "Invalid source ID"
+
+
+def test_create_source_success():
+    workspace_id = "my_workspace_id"
+    expected_response = {
+        "sourceId": "1",
+        "name": "Example Source 1",
+        "sourcedef_id": "1",
+        "config": {"test": "test"},
+    }
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 200
+        mock_post.return_value.json.return_value = expected_response
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+
+        result = create_source(workspace_id, "Example Source 1", "1", {"test": "test"})
+        assert result == expected_response
+        assert isinstance(result, dict)
+
+
+def test_create_source_failure():
+    workspace_id = "my_workspace_id"
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 500
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {"error": "Invalid request data"}
+        with pytest.raises(HttpError) as excinfo:
+            create_source(workspace_id, "Example Source 1", "1", {"test": "test"})
+        assert excinfo.value.status_code == 500
+        assert str(excinfo.value) == "failed to create source"
+
+
+def test_create_source_with_invalid_workspace_id():
+    with pytest.raises(HttpError) as excinfo:
+        create_source(123, "Example Source 1", "1", {"test": "test"})
+    assert str(excinfo.value) == "workspace_id must be a string"
+
+
+def test_create_source_with_invalid_source_name():
+    with pytest.raises(HttpError) as excinfo:
+        create_source("test", 123, "1", {"test": "test"})
+    assert str(excinfo.value) == "name must be a string"
+
+
+def test_create_source_with_invalid_sourcedef_id():
+    with pytest.raises(HttpError) as excinfo:
+        create_source("test", "Example Source 1", 123, {"test": "test"})
+    assert str(excinfo.value) == "sourcedef_id must be a string"
+
+
+def test_create_source_with_invalid_config():
+    with pytest.raises(HttpError) as excinfo:
+        create_source("test", "test", "test", 123)
+    assert str(excinfo.value) == "config must be a dictionary"
+
+
+def test_update_source_success():
+    name = "source"
+    source_id = "1"
+    sourcedef_id = "1"
+    expected_response = {
+        "sourceId": "1",
+        "name": "Example Source 1",
+        "config": {"test": "test"},
+        "sourcedef_id": "1",
+    }
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 200
+        mock_post.return_value.json.return_value = expected_response
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+
+        result = update_source(source_id, name, {"test": "test"}, sourcedef_id)
+        assert result == expected_response
+        assert isinstance(result, dict)
+
+
+def test_update_source_failure():
+    name = "source"
+    source_id = "1"
+    sourcedef_id = "1"
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 500
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+        mock_post.return_value.json.return_value = {"error": "Invalid request data"}
+        with pytest.raises(HttpError) as excinfo:
+            update_source(source_id, name, {"test": "test"}, sourcedef_id)
+        assert excinfo.value.status_code == 500
+        assert str(excinfo.value) == "failed to update source"
+
+
+def test_update_source_with_invalid_name():
+    with pytest.raises(HttpError) as excinfo:
+        update_source("1", 123, {"test": "test"}, "1")
+    assert str(excinfo.value) == "name must be a string"
+
+
+def test_update_source_with_invalid_source_id():
+    with pytest.raises(HttpError) as excinfo:
+        update_source(123, "test", {"test": "test"}, "1")
+    assert str(excinfo.value) == "source_id must be a string"
+
+
+def test_update_source_with_invalid_config():
+    with pytest.raises(HttpError) as excinfo:
+        update_source("test", "test", 123, "1")
+    assert str(excinfo.value) == "config must be a dictionary"
+
+
+def test_update_source_with_invalid_sourcedef_id():
+    with pytest.raises(HttpError) as excinfo:
+        update_source("test", "test", {"test": "test"}, 123)
+    assert str(excinfo.value) == "sourcedef_id must be a string"
+
+
+def test_check_source_connection_success():
+    workspace_id = "my_workspace_id"
+    data = AirbyteSourceCreate(
+        name="my_source_name",
+        sourceDefId="my_sourcedef_id",
+        config={"key": "value"},
+    )
+    expected_response = {"status": "succeeded"}
+
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 200
+        mock_post.return_value.json.return_value = expected_response
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+
+        result = check_source_connection(workspace_id, data)
+        assert result == expected_response
+        assert isinstance(result, dict)
+
+
+def test_check_source_connection_failure():
+    workspace_id = "my_workspace_id"
+    data = AirbyteSourceCreate(
+        name="my_source_name",
+        sourceDefId="my_sourcedef_id",
+        config={"key": "value"},
+    )
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_response = requests.Response()
+        mock_response.status_code = 500
+        mock_response._content = b'{"error": "failed to check source connection"}'
+        mock_post.return_value = mock_response
+        with pytest.raises(HttpError) as excinfo:
+            result = check_source_connection(workspace_id, data)
+            assert result is None
+            assert excinfo.value.status_code == 500
+            assert str(excinfo.value) == "failed to check source connection"
+
+
+def test_check_source_connection_with_invalid_workspace_id():
+    workspace_id = 123
+    data = AirbyteSourceCreate(
+        name="my_source_name",
+        sourceDefId="my_sourcedef_id",
+        config={"key": "value"},
+    )
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 500
+        mock_post.return_value.json.return_value = {"error": "Invalid request data"}
+        with pytest.raises(HttpError) as excinfo:
+            check_source_connection(workspace_id, data)
+        assert str(excinfo.value) == "workspace_id must be a string"
+
+
+def test_get_source_schema_catalog_success():
+    workspace_id = "my_workspace_id"
+    source_id = "my_source_id"
+    expected_response = {"catalog": "catalog"}
+
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 200
+        mock_post.return_value.json.return_value = expected_response
+        mock_post.return_value.headers = {"Content-Type": "application/json"}
+
+        result = get_source_schema_catalog(workspace_id, source_id)
+        assert result == expected_response
+        assert isinstance(result, dict)
+
+
+def test_get_source_schema_catalog_failure():
+    workspace_id = "my_workspace_id"
+    source_id = "my_source_id"
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_response = requests.Response()
+        mock_response.status_code = 500
+        mock_response._content = b'{"error": "failed to get source schema catalogs"}'
+        mock_post.return_value = mock_response
+        with pytest.raises(HttpError) as excinfo:
+            get_source_schema_catalog(workspace_id, source_id)
+            assert excinfo.value.status_code == 500
+            assert str(excinfo.value) == "failed to get source schema catalogs"
+
+
+def test_get_source_schema_catalog_with_invalid_workspace_id():
+    workspace_id = 123
+    source_id = "my_source_id"
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 500
+        mock_post.return_value.json.return_value = {"error": "Invalid request data"}
+        with pytest.raises(HttpError) as excinfo:
+            get_source_schema_catalog(workspace_id, source_id)
+        assert str(excinfo.value) == "workspace_id must be a string"
+
+
+def test_get_source_schema_catalog_with_invalid_source_id():
+    workspace_id = "my_workspace_id"
+    source_id = 123
+    with patch("ddpui.ddpairbyte.airbyte_service.requests.post") as mock_post:
+        mock_post.return_value.status_code = 500
+        mock_post.return_value.json.return_value = {"error": "Invalid request data"}
+        with pytest.raises(HttpError) as excinfo:
+            get_source_schema_catalog(workspace_id, source_id)
+        assert str(excinfo.value) == "source_id must be a string"
